refactor: rename useIsMobile hook file
