+++ components/hooks/useIsMobile.ts
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react'
+import { useEffect } from 'react'
 
 export const useIsMobile = (width: number, setWidth: any) => {
 	useEffect(() => {
